PowerStatus: Use BNumberFormat for percentage value
This provides automatic percentage formatting according to the locale
setting.

Reviewed-on: https://review.haiku-os.org/c/haiku/+/6209
Tested-by: Automation <automation@haiku-os.org>

+++ src/apps/powerstatus/PowerStatusView.h
@@ -64,7 +64,7 @@ protected:
 
 			battery_info	fBatteryInfo;
 
-			int32			fPercent;
+			double			fPercent;
 			time_t			fTimeLeft;
 			bool			fOnline;
 			bool			fHasBattery;
